CRITICAL FIX: Initialize prime_encoding and fix embedding storage type
DEPTH-13 ANALYSIS REVEALED CRITICAL BUGS:

Bug #1: Uninitialized prime_encoding (CRITICAL)
- Problem: token->prime_encoding was NEVER initialized in cllm_create.c
- Left as 0 from calloc(), causing L_lattice(0, ...) to compute near-zero values
- Result: ALL embeddings became 0.000000 (Mean=0, StdDev=0)
- Fix: Initialize with crystalline_get_nth_prime(i)

Bug #2: Float cast in double* storage (CRITICAL)
- Problem: embeddings[i] = (float)normalized cast double to float
- embeddings is double* but code was casting to float before storing
- Result: Precision loss (53-bit → 24-bit mantissa)
- Fix: Remove cast, store double directly

Bug #3: Symmetry group always zero (HIGH)
- Problem: All tokens assigned symmetry_group = 0
- Violates 12-fold symmetry principle
- Result: No symmetry diversity, reduced embedding quality
- Fix: Distribute across 12 groups with i % 12

Files Modified:
- src/ai/cllm_create.c: Initialize prime_encoding and symmetry_group
- src/ai/cllm_lattice_embeddings.c: Remove float cast in embedding storage

Impact:
- Embeddings will now have non-zero values
- Full 64-bit precision maintained
- 12-fold symmetry enforced
- Inference should work correctly

Build Status:
- Zero compilation errors
- Zero warnings
- All libraries built successfully

This fixes the root cause of 'Generation failed' error.
User's trained model had all-zero embeddings due to uninitialized prime_encoding.
New models will have properly initialized embeddings.

+++ src/ai/cllm_lattice_embeddings.c
@@ -84,7 +84,8 @@ void cllm_embeddings_init_lattice(CLLMModel* model) {
             if (normalized < -1.0) normalized = -1.0;
             
             // Store in embedding matrix
-            embeddings[token_id * embedding_dim + dim] = (float)normalized;
+            // CRITICAL FIX: Don't cast to float - embeddings is double*
+            embeddings[token_id * embedding_dim + dim] = normalized;
             
             // Update statistics
             sum += normalized;
